remove QuestionDetails component, WIP make PATCH more direct, without the need to click Save while creating forms
